Séparation des paramètres en 2 écrans différents maj notice
