+++ app/src/main/res/layout/fragment_home.xml
@@ -1,6 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
     android:orientation="horizontal"
     android:layout_width="match_parent"
     android:layout_height="match_parent">
@@ -29,6 +30,75 @@
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toBottomOf="@+id/recycler_view_buttons_one" />
+
+    <Button
+        android:id="@+id/exit_Button"
+        android:layout_width="wrap_content"
+        android:layout_height="60dp"
+        style="@style/default_text_style"
+        android:text="@string/exit"
+        android:layout_marginEnd="5dp"
+        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintEnd_toEndOf="parent" />
+
+    <Button
+        android:id="@+id/parameters_button"
+        android:layout_width="wrap_content"
+        android:layout_height="40dp"
+        android:fontFamily="@font/comic_neue_bold"
+        android:text="@string/app_parameters"
+        android:textSize="17sp"
+        android:layout_marginStart="5dp"
+        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintStart_toStartOf="parent" />
+
+    <LinearLayout
+        android:id="@+id/password_layout"
+        android:layout_width="150dp"
+        android:layout_height="160dp"
+        android:background="@drawable/layout_password"
+        android:elevation="5dp"
+        android:gravity="center_horizontal"
+        android:orientation="vertical"
+        android:padding="10dp"
+        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent">
+
+        <androidx.constraintlayout.utils.widget.MotionLabel
+            android:id="@+id/motionLabel"
+            style="@style/default_text_style"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:text="@string/password"
+            android:textColor="@color/white" />
+
+        <EditText
+            android:id="@+id/password_text_edit"
+            android:layout_width="100dp"
+            android:layout_height="50dp"
+            android:layout_marginTop="10dp"
+            android:layout_marginBottom="10dp"
+            android:background="@color/middle_blue"
+            android:inputType="none"
+            android:textColor="@color/white"
+            android:textCursorDrawable="@color/white"
+            android:textSize="30sp"
+            tools:ignore="TextFields,SpeakableTextPresentCheck" />
+
+        <Button
+            android:id="@+id/password_valid_button"
+            style="@style/default_text_style"
+            android:layout_width="80dp"
+            android:layout_height="50dp"
+            android:text="OK"
+            app:strokeColor="#3F9AA6"
+            app:strokeWidth="2sp" />
+
+
+    </LinearLayout>
+
 
 
 </androidx.constraintlayout.widget.ConstraintLayout>
